speech(voice-loop): narrow macos backends gates (#137)
Extract `build_local_backends_macos_native` and
`build_local_backends_macos_native_26` into a sibling module
`backends_macos.rs` gated on `target_os = "macos" + cpal +
any(macos-native, macos-native-26)`. Move shared types
`LocalBackends`/`ChannelStt` to a new `backends_common.rs`
gated only on `cpal`. The whisper+piper-specific
`build_local_backends` stays in `backends.rs` with the full
`silero+whisper+piper+cpal` gate.

The original `voice_loop/mod.rs` re-exported both macos
builders behind the full `silero+whisper+piper+cpal` set,
even though `build_local_backends_macos_native` only uses
silero+cpal+macos-native and `build_local_backends_macos_native_26`
only uses cpal+macos-native-26. The over-gating was a
consequence of `backends.rs` hosting all three builders;
splitting the file lets each re-export carry the narrowest
gate that actually reflects its body.

Practical impact today: none — every user enters via
`--features primer-cli/speech` which pulls all four
features and the parent `pub mod voice_loop;` in `lib.rs`
is still gated on `feature = "voice-loop"` = silero+whisper+piper+cpal.
A future PR can loosen that parent gate without touching
the inner gates, which now correctly reflect each
function's true dependencies.
